+++ changes.md
@@ -9,10 +9,11 @@
 * Fix ShieldShape bug where the base height would break in a way that made the shape not actually symetrical.
 * Shieldwall set added support for setting base_height to all parts.
 * Shieldwall added Magnets
-* Integrated Magnets into 
+* Integrated Magnets into:
   * StraightBasic
   * Straight
-  * EndCap
+  * EndCaps
+  * CornerConnector
 
 ## 3.1.0
 * Added shieldwall Set